requirements.txt
# Actual Dependencies
-e .
pynacl>0.0.3

# Test Dependencies
nose==1.3.6
coverage>=3.6,<3.99
mock>=1.0.1,<1.0.99
sphinx>=1.2.3
python-coveralls>=2.4.2
hypothesis==1.0.0
bumpversion
tox
yanc
